requirements.txt
# Local
-e .

click==8.0.1
dvc[s3]==2.3.0
dvclive==0.0.3
jupyter==1.0.0
pillow==8.2.0

# Torch
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio==0.9.0 